move preprints migration to the scripts folder
